remove an extra folder, moved project files
