+++ core/db_helper.py
@@ -2,7 +2,7 @@ from asyncio import current_task
 
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker, async_scoped_session, AsyncSession
 
-from pc_store.core.config import settings
+from core.config import settings
 
 
 class DataBaseHelper:
