--- pc_store/pyproject.toml
@@ -1,19 +1,0 @@
-[project]
-name = "pc-store"
-version = "0.1.0"
-description = ""
-authors = [
-    {name = "danil.bobryshov",email = "danil.bobryshov@mail.ru"}
-]
-readme = "README.md"
-requires-python = ">=3.9"
-dependencies = [
-]
-
-[tool.poetry]
-packages = [{include = "pc_store", from = "src"}]
-
-
-[build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
