+++ repositories/user_repository.py
@@ -4,7 +4,7 @@ READ
 UPDATE
 DELETE
 """
-from pc_store.models import User
+from models import User
 
 
 def create_user(user_in: User) -> dict:
